HttpRequestGeneratorClient/requirements.txt
requests
opencensus-ext-azure
opencensus-ext-logging
opencensus-ext-requests